Move format-indepenent private utilities to manifest/common.go
separate from the format-specific files, and from the ~generic
public utilities in manifest.go.

Only moves code, should not change behavior.
