+++ manifest/manifest_test.go
@@ -5,7 +5,6 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/containers/image/v5/types"
 	"github.com/containers/libtrust"
 	digest "github.com/opencontainers/go-digest"
 	imgspecv1 "github.com/opencontainers/image-spec/specs-go/v1"
@@ -162,50 +161,4 @@ func TestNormalizedMIMEType(t *testing.T) {
 		res := NormalizedMIMEType(c)
 		assert.Equal(t, DockerV2Schema1SignedMediaType, res, c)
 	}
-}
-
-func TestLayerInfosToStrings(t *testing.T) {
-	strings := layerInfosToStrings([]LayerInfo{})
-	assert.Equal(t, []string{}, strings)
-
-	strings = layerInfosToStrings([]LayerInfo{
-		{
-			BlobInfo: types.BlobInfo{
-				MediaType: "application/vnd.docker.image.rootfs.diff.tar.gzip",
-				Digest:    "sha256:a3ed95caeb02ffe68cdd9fd84406680ae93d633cb16422d00e8a7c22955b46d4",
-				Size:      32,
-			},
-			EmptyLayer: true,
-		},
-		{
-			BlobInfo: types.BlobInfo{
-				MediaType: "application/vnd.docker.image.rootfs.diff.tar.gzip",
-				Digest:    "sha256:bbd6b22eb11afce63cc76f6bc41042d99f10d6024c96b655dafba930b8d25909",
-				Size:      8841833,
-			},
-			EmptyLayer: false,
-		},
-		{
-			BlobInfo: types.BlobInfo{
-				MediaType: "application/vnd.docker.image.rootfs.diff.tar.gzip",
-				Digest:    "sha256:960e52ecf8200cbd84e70eb2ad8678f4367e50d14357021872c10fa3fc5935fa",
-				Size:      291,
-			},
-			EmptyLayer: false,
-		},
-		{
-			BlobInfo: types.BlobInfo{
-				MediaType: "application/vnd.docker.image.rootfs.diff.tar.gzip",
-				Digest:    "sha256:a3ed95caeb02ffe68cdd9fd84406680ae93d633cb16422d00e8a7c22955b46d4",
-				Size:      32,
-			},
-			EmptyLayer: true,
-		},
-	})
-	assert.Equal(t, []string{
-		"sha256:a3ed95caeb02ffe68cdd9fd84406680ae93d633cb16422d00e8a7c22955b46d4",
-		"sha256:bbd6b22eb11afce63cc76f6bc41042d99f10d6024c96b655dafba930b8d25909",
-		"sha256:960e52ecf8200cbd84e70eb2ad8678f4367e50d14357021872c10fa3fc5935fa",
-		"sha256:a3ed95caeb02ffe68cdd9fd84406680ae93d633cb16422d00e8a7c22955b46d4",
-	}, strings)
 }
